pyproject.toml
[tool.poetry]
name = "django-poetry-rest-api"
version = "0.1.0"
description = ""
authors = ["André Kovac <me@andrekovac.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.942"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
